+++ src/main/java/Rules/PowerRule.java
@@ -26,7 +26,7 @@ public class PowerRule extends DerivationRule {
     }
 
     @Override
-    public int getResult(List<Integer> dims) {
+    public double getResult(List<Integer> dims) {
         return (int)Math.pow((double)this.terms.get(0).evaluate(dims), (double)this.terms.get(1).evaluate(dims));
     }
