+++ src/main/java/TrigFunctions/Trigenometric/SecRule.java
@@ -19,7 +19,7 @@ public class SecRule extends DerivationRule {
     }
 
     @Override
-    public int getResult(List<Integer> dims) {
+    public double getResult(List<Integer> dims) {
         return 0;
     }
 }
